Chord Organ 2
This changes the knobs on the face of the unit from Chord and Root, to Glide and Chorus. Chord and Root must be run entirely from CV in this version of the software, including whatever offsets you need to get the root octave correct. In return for giving up the knob adjustments, now you can manipulate glide and chorus speed on the fly. Chorus of zero means it switches out of Stacked mode, or you can have incredibly slow chorusing, or octave/pitched-up effects. Glide is always engaged but can be too fast to hear as a portamento: also, the technique for applying it is frequency chasing rather than linear ramping so the result is more 303-like in style.